requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
psycopg2==2.8.5
bcrypt==3.2.0
Flask-Cors==3.0.9
numpy==1.19.3
pandas==1.1.4
XlsxWriter==1.3.7
xlwt==1.3.0
marshmallow==3.7.1
cloudinary==1.24.0
PyJWT==2.0.1